Modified tests and added new cases
