+++ tests/data/wire/multi-config/ext_conf_multi_config_no_dependencies.xml
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Extensions version="1.0.0.0" goalStateIncarnation="4"><GuestAgentExtension xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
+  <GAFamilies>
+      <GAFamily>
+          <Name>Prod</Name>
+          <Uris>
+              <Uri>http://mock-goal-state/manifest_of_ga.xml</Uri>
+          </Uris>
+      </GAFamily>
+      <GAFamily>
+          <Name>Test</Name>
+          <Uris>
+              <Uri>http://mock-goal-state/manifest_of_ga.xml</Uri>
+          </Uris>
+      </GAFamily>
+  </GAFamilies>
+  <Location>eastus</Location>
+  <ServiceName>CRP</ServiceName>
+</GuestAgentExtension>
+<RequiredFeatures>
+  <RequiredFeature>
+    <Name>MultipleExtensionsPerHandler</Name>
+  </RequiredFeature>
+</RequiredFeatures>
+<StatusUploadBlob statusBlobType="BlockBlob">https://test.blob.core.windows.net/$system/lrwinmcdn_0.0f3bfecf-f14f-4c7d-8275-9dee7310fe8c.status?sv=2018-03-28&amp;sr=b&amp;sk=system-1&amp;sig=1%2b%2f4nL3kZJyUb7EKxSVGQ%2fHLpXBZxCU8Zo4diPFPv5o%3d&amp;se=9999-01-01T00%3a00%3a00Z&amp;sp=w</StatusUploadBlob>
+<Plugins>
+  <Plugin name="OSTCExtensions.ExampleHandlerLinux" version="1.0.0" location="http://mock-goal-state/rdfepirv2hknprdstr03.blob.core.windows.net/b01058962be54ceca550a390fa5ff064/Microsoft.OSTCExtensions_ExampleHandlerLinux_asiaeast_manifest.xml" config="" state="enabled" autoUpgrade="false" failoverlocation="http://mock-goal-state/rdfepirv2hknprdstr04.blob.core.windows.net/b01058962be54ceca550a390fa5ff064/Microsoft.OSTCExtensions_ExampleHandlerLinux_asiaeast_manifest.xml" runAsStartupTask="false" isJson="true" />
+  <Plugin name="Microsoft.Powershell.ExampleExtension" version="1.0.0" location="https://mock-goal-state/zrdfepirv2sn3prdstr06.blob.core.windows.net/3055d769a0ed485d8c5ea8a7cd24073e/Microsoft.Powershell_ExampleExtension_ussouth_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="https://mock-goal-state/zrdfepirv2sn4prdstr02a.blob.core.windows.net/3055d769a0ed485d8c5ea8a7cd24073e/Microsoft.Powershell_ExampleExtension_ussouth_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="false"/>
+</Plugins>
+<PluginSettings>
+  <Plugin name="OSTCExtensions.ExampleHandlerLinux" version="1.0.0" >
+    <ExtensionRuntimeSettings seqNo="1" name="firstExtension" state="enabled">{
+  "runtimeSettings": [
+    {
+      "handlerSettings": {
+        "publicSettings": {"source":{"script":"Write-Host First: Hello World 1234!"},"parameters":[{"name":"extensionName","value":"firstRunCommand"}],"timeoutInSeconds":120}
+      }
+    }
+  ]
+}</ExtensionRuntimeSettings>
+    <ExtensionRuntimeSettings seqNo="2" name="secondExtension" state="enabled">{
+  "runtimeSettings": [
+    {
+      "handlerSettings": {
+        "publicSettings": {"source":{"script":"Write-Host First: Hello World 1234!"},"parameters":[{"name":"extensionName","value":"secondRunCommand"}],"timeoutInSeconds":120}
+      }
+    }
+  ]
+}</ExtensionRuntimeSettings>
+    <ExtensionRuntimeSettings seqNo="3" name="thirdExtension" state="enabled">{
+  "runtimeSettings": [
+    {
+      "handlerSettings": {
+        "publicSettings": {"source":{"script":"Write-Host Third: Hello World 3!"},"parameters":[{"name":"extensionName","value":"thirdRunCommand"}],"timeoutInSeconds":120}
+      }
+    }
+  ]
+}</ExtensionRuntimeSettings>
+  </Plugin>
+  <Plugin name="Microsoft.Powershell.ExampleExtension" version="1.0.0">
+    <RuntimeSettings seqNo="9">{
+  "runtimeSettings": [
+    {
+      "handlerSettings": {
+        "publicSettings": {"fileUris":["https://test.blob.core.windows.net/windowsagent/VerifyAgentRunning.ps1"],"commandToExecute":"echo Hello 1234"}
+      }
+    }
+  ]
+}</RuntimeSettings>
+  </Plugin>
+</PluginSettings>
+<InVMArtifactsProfileBlob>https://test.blob.core.windows.net/$system/lrwinmcdn_0.0f3bfecf-f14f-4c7d-8275-9dee7310fe8c.vmSettings?sv=2018-03-28&amp;sr=b&amp;sk=system-1&amp;sig=8YHwmibhasT0r9MZgL09QmFwL7ZV%2bg%2b49QP5Zwe4ksY%3d&amp;se=9999-01-01T00%3a00%3a00Z&amp;sp=r</InVMArtifactsProfileBlob>
+<InVMGoalStateMetaData inSvdSeqNo="3" createdOnTicks="637415129555522297" />
+<HealthProbe id="0f3bfecf-f14f-4c7d-8275-9dee7310fe8c" type="loadbalancer" port="80" />
+<HealthStatusBlobUri>https://test.blob.core.windows.net/$system/lrwinmcdn_0.0f3bfecf-f14f-4c7d-8275-9dee7310fe8c.vmHealth?sv=2018-03-28&amp;sr=b&amp;sk=system-1&amp;sig=DQSxfPRZEoGBGIFl%2f4bFZ0LM9RNr9DbUEmmtkiQkWkE%3d&amp;se=9999-01-01T00%3a00%3a00Z&amp;sp=rw</HealthStatusBlobUri></Extensions>
