+++ tests/data/wire/multi-config/ext_conf_with_multi_config.xml
@@ -24,14 +24,14 @@
 </RequiredFeatures>
 <StatusUploadBlob statusBlobType="PageBlob">https://test.blob.core.windows.net/$system/lrwinmcdn_0.0f3bfecf-f14f-4c7d-8275-9dee7310fe8c.status?sv=2018-03-28&amp;sr=b&amp;sk=system-1&amp;sig=1%2b%2f4nL3kZJyUb7EKxSVGQ%2fHLpXBZxCU8Zo4diPFPv5o%3d&amp;se=9999-01-01T00%3a00%3a00Z&amp;sp=w</StatusUploadBlob>
 <Plugins>
-  <Plugin name="Microsoft.CPlat.Core.RunCommandHandlerWindows" version="2.0.2" location="http://mock-goal-state/rdfepirv2bl3prdstr01.blob.core.windows.net/f72653efd9e349ed9842c8b99e4c1712/Microsoft.CPlat.Core_RunCommandHandlerWindows_useast_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="http://mock-goal-state/zrdfepirv2bl5prdstr06.blob.core.windows.net/f72653efd9e349ed9842c8b99e4c1712/Microsoft.CPlat.Core_RunCommandHandlerWindows_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
+  <Plugin name="Microsoft.CPlat.Core.RunCommandHandlerWindows" version="2.3.0" location="http://mock-goal-state/rdfepirv2bl3prdstr01.blob.core.windows.net/f72653efd9e349ed9842c8b99e4c1712/Microsoft.CPlat.Core_RunCommandHandlerWindows_useast_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="http://mock-goal-state/zrdfepirv2bl5prdstr06.blob.core.windows.net/f72653efd9e349ed9842c8b99e4c1712/Microsoft.CPlat.Core_RunCommandHandlerWindows_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
   <Plugin name="Microsoft.Azure.Security.AntimalwareSignature.AntimalwareConfiguration" version="2.0.53.0" location="http://mock-goal-state/zrdfepirv2bl6prdstr05a.blob.core.windows.net/4eb6e001f5e54cb9b17d94fd6ffad9a7/Microsoft.Azure.Security.AntimalwareSignature_AntimalwareConfiguration_useast_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="http://mock-goal-state/rdfepirv2bl3prdstr03.blob.core.windows.net/4eb6e001f5e54cb9b17d94fd6ffad9a7/Microsoft.Azure.Security.AntimalwareSignature_AntimalwareConfiguration_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
   <Plugin name="Microsoft.Azure.Geneva.GenevaMonitoring" version="2.20.0.1" location="http://mock-goal-state/rdfepirv2bl2prdstr02.blob.core.windows.net/dfe9924e0f8c4d6d83e1d19f1e8238fc/Microsoft.Azure.Geneva_GenevaMonitoring_useast_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="http://mock-goal-state/zrdfepirv2bl5prdstr06.blob.core.windows.net/dfe9924e0f8c4d6d83e1d19f1e8238fc/Microsoft.Azure.Geneva_GenevaMonitoring_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
   <Plugin name="Microsoft.Compute.CustomScriptExtension" version="1.10.9" location="http://mock-goal-state/zrdfepirv2bl6prdstr05a.blob.core.windows.net/bfd5c281a7dc4e4b84381eb0b47e3aaf/Microsoft.Compute_CustomScriptExtension_useast_manifest.xml" state="enabled" autoUpgrade="false" failoverlocation="http://mock-goal-state/rdfepirv2bl2prdstr02.blob.core.windows.net/bfd5c281a7dc4e4b84381eb0b47e3aaf/Microsoft.Compute_CustomScriptExtension_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
   <Plugin name="Microsoft.Compute.VMAccessAgent" version="2.4.7" location="http://mock-goal-state/rdfepirv2bl3prdstr01.blob.core.windows.net/bfd5c281a7dc4e4b84381eb0b47e3aaf/Microsoft.Compute_VMAccessAgent_useast_manifest.xml" state="enabled" autoUpgrade="true" failoverlocation="http://mock-goal-state/zrdfepirv2bl5prdstr06.blob.core.windows.net/bfd5c281a7dc4e4b84381eb0b47e3aaf/Microsoft.Compute_VMAccessAgent_useast_manifest.xml" runAsStartupTask="false" isJson="true" useExactVersion="true" />
 </Plugins>
 <PluginSettings>
-  <Plugin name="Microsoft.CPlat.Core.RunCommandHandlerWindows" version="2.0.2">
+  <Plugin name="Microsoft.CPlat.Core.RunCommandHandlerWindows" version="2.3.0">
     <DependsOn dependencyLevel="3" name="secondRunCommand">
       <DependsOnExtension extension="firstRunCommand" handler="Microsoft.CPlat.Core.RunCommandHandlerWindows" />
       <DependsOnExtension handler="Microsoft.Compute.CustomScriptExtension" />
